Fixed https://github.com/labstack/echo/issues/507

+++ bytes/bytes_test.go
@@ -36,56 +36,56 @@ func TestBytesParse(t *testing.T) {
 	// B
 	b, err := Parse("515B")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 515, b)
+		assert.Equal(t, int64(515), b)
 	}
 
 	// KB
 	b, err = Parse("12KB")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 12288, b)
+		assert.Equal(t, int64(12288), b)
 	}
 	b, err = Parse("12K")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 12288, b)
+		assert.Equal(t, int64(12288), b)
 	}
 
 	// MB
 	b, err = Parse("2MB")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 2097152, b)
+		assert.Equal(t, int64(2097152), b)
 	}
 	b, err = Parse("2M")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 2097152, b)
+		assert.Equal(t, int64(2097152), b)
 	}
 
 	// GB
 	b, err = Parse("6GB")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 6442450944, b)
+		assert.Equal(t, int64(6442450944), b)
 	}
 	b, err = Parse("6G")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 6442450944, b)
+		assert.Equal(t, int64(6442450944), b)
 	}
 
 	// TB
 	b, err = Parse("5TB")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 5497558138880, b)
+		assert.Equal(t, int64(5497558138880), b)
 	}
 	b, err = Parse("5T")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 5497558138880, b)
+		assert.Equal(t, int64(5497558138880), b)
 	}
 
 	// PB
 	b, err = Parse("9PB")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 10133099161583616, b)
+		assert.Equal(t, int64(10133099161583616), b)
 	}
 	b, err = Parse("9P")
 	if assert.NoError(t, err) {
-		assert.Equal(t, 10133099161583616, b)
+		assert.Equal(t, int64(10133099161583616), b)
 	}
 }
